% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2geno.R
\name{counts2geno}
\alias{counts2geno}
\title{Convert count data to genotypes}
\usage{
counts2geno(count.data, pop.size = 10)
}
\arguments{
\item{count.data}{An object created via the function read.dart.counts
[required].}

\item{pop.size}{Number of individuals to be simulated [default = 10].}
}
\value{
A genlight object.
}
\description{
Convert count data to genotypes
}
