% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.assignment.R
\name{dart.assignment}
\alias{dart.assignment}
\title{Population assignment probabilities}
\usage{
dart.assignment(ref, unknown)
}
\arguments{
\item{ref}{Genlight object containing the reference data [required].}

\item{unknown}{Genlight object containing the individual to be assigned to a
reference [required].}
}
\value{
A \code{data.frame} consisting of assignment probabilities for each
 population.
}
\description{
This function takes one sample and estimates its probability of coming from
a reference
}
\details{
This function is a re-implementation of the function multilocus_assignment
 from package gstudio.
 Description of the method used in this function can be found at:
https://dyerlab.github.io/applied_population_genetics/population-assignment.html
}
