% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartDistanceMatrix'.R
\name{dartDistanceMatrix}
\alias{dartDistanceMatrix}
\title{dartDistanceMatrix}
\usage{
dartDistanceMatrix(
  genotypic_data,
  sampleWiseAnalysis = F,
  ncores = parallel::detectCores()
)
}
\description{
Calls the DArT implementation of distance matrix
}
