% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{dartVarietalIDShiny}
\alias{dartVarietalIDShiny}
\title{dartVarietalIDShiny}
\usage{
dartVarietalIDShiny()
}
\description{
dartVarietalIDShiny
}
\examples{
\dontrun{
browseURL(system.file("extdata",package = "dartVarietalID")) # Open folder of example files
dartVarietalIDShiny()
}
}
