% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dart.counts.R
\name{read.dart.counts}
\alias{read.dart.counts}
\title{Reads count data and info file}
\usage{
read.dart.counts(counts.file, info.file)
}
\arguments{
\item{counts.file}{csv file containing the count data [required].}

\item{info.file}{csv file containing samples information [required].}
}
\value{
A list with the following elements: counts, loc_metrics, ind_metrics
}
\description{
Reads count data and info file
}
