% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_main.R
\name{runSampleAnalysis}
\alias{runSampleAnalysis}
\title{Main script}
\usage{
runSampleAnalysis(
  counts.file,
  info.file,
  ncores = parallel::detectCores() - 1,
  pop.size = 10
)
}
\arguments{
\item{counts.file}{File with count data [required].}

\item{info.file}{File with information of samples and references[required].}

\item{ncores}{Number of cores to run analysis
[default parallel::detectCores() - 1].}

\item{pop.size}{Number individuals to simulate [default 10].}
}
\value{
A list with the following elements:

res_summary. A data frame containing the best reference match for each sample
and its probability.

res_full. A list of data frames with the probability match for each sample in
 the dataset.

gl.references. A genlight object containing the references.

gl.samples. A genlight object containing the samples.
}
\description{
Main script
}
\details{
Main script
}
\examples{

filename <- system.file('extdata','SEQ_SNPs_counts_0_Target.csv',
package='dartVarietalID')
info <- system.file('extdata','InfoFile_corrected.csv',
package='dartVarietalID')
ID_res <- runSampleAnalysis(counts.file = filename, info.file = info,
ncores = 1, pop.size = 10)

}
\author{
Luis Mijangos
}
