% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.assignment.R
\name{dart.assignment}
\alias{dart.assignment}
\title{Population assignment probabilities}
\usage{
dart.assignment(ref, unknown)
}
\arguments{
\item{ref}{A list of genlight objects each containing one reference
[required].}

\item{unknown}{A genlight object containing the sample to be assigned to a
reference [required].}
}
\value{
A \code{data.frame} consisting of assignment probabilities for each
 population.
}
\description{
This function takes one sample and estimates its probability of coming from
a reference
}
