% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartDistanceMatrix.R
\name{dartDistanceMatrix}
\alias{dartDistanceMatrix}
\title{dartDistanceMatrix}
\usage{
dartDistanceMatrix(
  genotypic_data,
  sampleWiseAnalysis = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{genotypic_data}{description}

\item{sampleWiseAnalysis}{description}

\item{ncores}{description
Calls the DArT implementation of distance matrix}
}
\description{
dartDistanceMatrix
}
